+++ core/deferredGraphics/renderStages/source/lighting.cpp
@@ -128,7 +128,7 @@ void graphics::updateLightingDescriptorSets(camera* cameraObject)
             descriptorWrites.push_back(VkWriteDescriptorSet{});
                 descriptorWrites.back().sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
                 descriptorWrites.back().dstSet = lighting.DescriptorSets[i];
-                descriptorWrites.back().dstBinding = descriptorWrites.size() - 1;
+                descriptorWrites.back().dstBinding = static_cast<uint32_t>(descriptorWrites.size() - 1);
                 descriptorWrites.back().dstArrayElement = 0;
                 descriptorWrites.back().descriptorType = VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT;
                 descriptorWrites.back().descriptorCount = 1;
@@ -137,7 +137,7 @@ void graphics::updateLightingDescriptorSets(camera* cameraObject)
         descriptorWrites.push_back(VkWriteDescriptorSet{});
             descriptorWrites.back().sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
             descriptorWrites.back().dstSet = lighting.DescriptorSets[i];
-            descriptorWrites.back().dstBinding = descriptorWrites.size() - 1;
+            descriptorWrites.back().dstBinding = static_cast<uint32_t>(descriptorWrites.size() - 1);
             descriptorWrites.back().dstArrayElement = 0;
             descriptorWrites.back().descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
             descriptorWrites.back().descriptorCount = 1;
